共通処理を整理
