+++ frontend/src/app/dividendPortfolio/page.tsx
@@ -4,6 +4,7 @@ import { useEffect, useState, useRef } from "react";
 import { Chart as ChartJS, CategoryScale, LinearScale, PointElement, LineElement, Title, Tooltip, Legend, Filler, ArcElement } from "chart.js";
 import { Chart } from "react-chartjs-2";
 import { getAuthHeaders } from "@/utils/auth";
+import { exportChartImage } from "@/utils/exportChartImage";
 
 ChartJS.register(CategoryScale, LinearScale, PointElement, LineElement, Title, Tooltip, Legend, Filler, ArcElement);
 
@@ -61,15 +62,7 @@ export default function DividendPortfolio() {
     };
 
     const handleExport = () => {
-        if (chartRef.current) {
-            const url = chartRef.current.toBase64Image();
-            const link = document.createElement("a");
-            link.href = url;
-            const now = new Date();
-            const timestamp = now.toISOString().replace(/[:-]/g, "").replace(/\..+/, "");
-            link.download = `dividendPortfolioChart_${timestamp}.png`;
-            link.click();
-        }
+        exportChartImage(chartRef, "dividendPortfolioChart");
     };
 
     if (error) return <p className="text-red-500">{error}</p>;
